+++ src/main/java/net/depoker/createjs/easeljs/client/geom/impl/RectangleImpl.java
@@ -1,3 +1,7 @@
+/*
+ * Copyright 2012-2015 Double Edged Poker
+ */
+
 package net.depoker.createjs.easeljs.client.geom.impl;
 
 import com.google.gwt.core.client.JavaScriptObject;
@@ -12,7 +16,7 @@ public class RectangleImpl extends JavaScriptObject {
 	public static native RectangleImpl create() /*-{
 		return new $wnd.createjs.Rectangle();
 	}-*/;
-	
+
 	/**
 	* Constructs a new Rectangle instance.
 	* Represents a rectangle as defined by the points (x,y) and (x+w,y+h).
@@ -21,43 +25,43 @@ public class RectangleImpl extends JavaScriptObject {
 	* @param w Width. Default is 0.
 	* @param h Height. Default is 0.
 	*/
-	public static native RectangleImpl create(float x, float y, float w, float h) /*-{
+	public static native RectangleImpl create(double x, double y, double w, double h) /*-{
 		return new $wnd.createjs.Rectangle(x, y, w, h);
 	}-*/;
-	
+
 	public final native RectangleImpl clone() /*-{
 		return this.clone();
 	}-*/;
 
-	public final native float getX() /*-{
+	public final native double getX() /*-{
 		return this.x;
 	}-*/;
-	
-	public final native void setX(float x) /*-{
+
+	public final native void setX(double x) /*-{
 		this.x = x;
 	}-*/;
-	
-	public final native float getY() /*-{
+
+	public final native double getY() /*-{
 		return this.y;
 	}-*/;
-	
-	public final native void setY(float y) /*-{
+
+	public final native void setY(double y) /*-{
 		this.y = y;
 	}-*/;
 
-	public final native float getWidth() /*-{
+	public final native double getWidth() /*-{
 		return this.width;
 	}-*/;
-	
-	public final native void setWidth(float w) /*-{
+
+	public final native void setWidth(double w) /*-{
 		this.width = w;
 	}-*/;
 
-	public final native float getHeight() /*-{
+	public final native double getHeight() /*-{
 		return this.height
 	}-*/;
-	
-	public final native void setHeight(float h) /*-{
+
+	public final native void setHeight(double h) /*-{
 		this.height = h;
 	}-*/;
 }
